Completed the TypeScript portion for now
I've filled in all the TypeScript portions of it. I now want to try and use it somewhere. I'm still not that good at Web Dev, but if I can improve it, that would be great. Want to learn a little bit of Bulma CSS Framework (very simple, but effective framework), then make something